ftp: a 550 response to SIZE is now an error
and returns CURLE_FTP_COULDNT_RETR_FILE. This is primarily interesting
for cases where CURLOPT_NOBODY is set as otherwise curl would not return
an error for this case.

This is how libcurl already acts on a 550 as a MDTM response (when
CURLOPT_FILETIME is set). If CURLOPT_NOBODY is not set, the error will
happen subsequently anyway since the RETR command will also fail with
it.

Reported-by: Tomas Berger
Fixes #5953